refactor into ABC and subclass
- single MigrationHelper for now. ABC supports
potential InitializeHelper.
